feat: Improve dropdown positioning and responsiveness
Adds a `triggerClassName` prop to `SmartDropdown` to allow for more flexible trigger styling. Also, introduces a double `requestAnimationFrame` in `SmartDropdown` to ensure correct positioning of the dropdown content on initial render, fixing a potential visual glitch. The `triggerClassName` is set to "h-auto" on various font and number format selectors to accommodate their content height.